core: Rename IdConflict.detect() to Conflict.detectOnIdentity()
So we can later add more conflict types.
